# reports/verify-61c0f29/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Microservice Code Studio",
    "description": "Microservice Code Studio - Autonomous Java 21 / Spring Boot 3.x Generator",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/orchestrator/sessions/{session_id}/overview": {
      "get": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Session Overview",
        "description": "Retrieves high-level project summary and milestone metrics for the Home View.",
        "operationId": "session_overview_api_v1_orchestrator_sessions__session_id__overview_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOverviewSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/sessions/{session_id}/lifecycle": {
      "get": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Session Lifecycle",
        "description": "Returns the full lifecycle state, phase progress, and next recommended action.",
        "operationId": "session_lifecycle_api_v1_orchestrator_sessions__session_id__lifecycle_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/sessions/{session_id}/transition": {
      "post": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Transition Session Phase",
        "description": "Validates prerequisites and transitions the session to a target lifecycle phase.",
        "operationId": "transition_session_phase_api_v1_orchestrator_sessions__session_id__transition_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhaseTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/sessions/{session_id}/invalidate": {
      "post": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Invalidate Downstream Phases",
        "description": "Marks downstream phases as OUTDATED after an upstream artifact modification.",
        "operationId": "invalidate_downstream_phases_api_v1_orchestrator_sessions__session_id__invalidate_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/pipeline/run": {
      "post": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Run Autopilot Pipeline",
        "description": "Initiates autonomous background Auto-Pilot pipeline execution.",
        "operationId": "run_autopilot_pipeline_api_v1_orchestrator_pipeline_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/pipeline/pause": {
      "post": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Pause Autopilot Pipeline",
        "description": "Pauses an active Auto-Pilot run and transitions to Guided Step-by-Step mode.",
        "operationId": "pause_autopilot_pipeline_api_v1_orchestrator_pipeline_pause_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/pipeline/{session_id}/pause": {
      "post": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Pause Autopilot Pipeline",
        "description": "Pauses an active Auto-Pilot run and transitions to Guided Step-by-Step mode.",
        "operationId": "pause_autopilot_pipeline_api_v1_orchestrator_pipeline__session_id__pause_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/pipeline/resume": {
      "post": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Resume Autopilot Pipeline",
        "description": "Resumes a paused Auto-Pilot run.",
        "operationId": "resume_autopilot_pipeline_api_v1_orchestrator_pipeline_resume_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/pipeline/{session_id}/resume": {
      "post": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Resume Autopilot Pipeline",
        "description": "Resumes a paused Auto-Pilot run.",
        "operationId": "resume_autopilot_pipeline_api_v1_orchestrator_pipeline__session_id__resume_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/pipeline/cancel": {
      "post": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Cancel Autopilot Pipeline",
        "description": "Cancels an active or paused Auto-Pilot run.",
        "operationId": "cancel_autopilot_pipeline_api_v1_orchestrator_pipeline_cancel_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/pipeline/{session_id}/cancel": {
      "post": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Cancel Autopilot Pipeline",
        "description": "Cancels an active or paused Auto-Pilot run.",
        "operationId": "cancel_autopilot_pipeline_api_v1_orchestrator_pipeline__session_id__cancel_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/pipeline/{session_id}/events": {
      "get": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Stream Pipeline Progress",
        "description": "Streams real-time progress events for an active pipeline via Server-Sent Events (SSE).",
        "operationId": "stream_pipeline_progress_api_v1_orchestrator_pipeline__session_id__events_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/sessions/{session_id}/export-bundle": {
      "get": {
        "tags": [
          "Lifecycle Orchestrator & Auto-Pilot"
        ],
        "summary": "Export Bundle Archive",
        "description": "Downloads all workspace artifacts in a unified ZIP archive.",
        "operationId": "export_bundle_archive_api_v1_orchestrator_sessions__session_id__export_bundle_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LifecycleState": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "currentPhase": {
            "$ref": "#/components/schemas/LifecyclePhase",
            "default": "INITIAL"
          },
          "completionPercentage": {
            "type": "number",
            "title": "Completionpercentage",
            "default": 0.0
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/PhaseState"
            },
            "type": "array",
            "title": "Phases"
          },
          "nextRecommendedAction": {
            "type": "string",
            "title": "Nextrecommendedaction",
            "default": "Ingresar o confirmar requisitos del microservicio"
          },
          "nextTargetPhase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LifecyclePhase"
              },
              {
                "type": "null"
              }
            ],
            "default": "REQUIREMENTS"
          },
          "canAdvance": {
            "type": "boolean",
            "title": "Canadvance",
            "default": true
          },
          "activeMode": {
            "$ref": "#/components/schemas/PipelineExecutionMode",
            "default": "GUIDED_STEP"
          },
          "isOutdated": {
            "type": "boolean",
            "title": "Isoutdated",
            "default": false
          },
          "pipelineStatus": {
            "$ref": "#/components/schemas/PipelineRunStatus",
            "default": "IDLE"
          }
        },
        "type": "object",
        "required": [
          "sessionId"
        ],
        "title": "LifecycleState"
      },
...
      "PhaseTransitionRequest": {
        "properties": {
          "targetPhase": {
            "$ref": "#/components/schemas/LifecyclePhase"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "apiKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "targetPhase"
        ],
        "title": "PhaseTransitionRequest"
      },
      "PipelineExecutionMode": {
        "type": "string",
        "enum": [
          "AUTO_PILOT",
          "GUIDED_STEP"
        ],
        "title": "PipelineExecutionMode"
      },
      "PipelineRunRequest": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "targetPhase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LifecyclePhase"
              },
              {
                "type": "null"
              }
            ],
            "default": "DEVOPS_DEPLOY"
          },
          "stopOnGate": {
            "type": "boolean",
            "title": "Stopongate",
            "default": true
          },
          "autoDeploy": {
            "type": "boolean",
            "title": "Autodeploy",
            "default": false
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "apiKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "sessionId"
        ],
        "title": "PipelineRunRequest"
      },
      "PipelineRunStatus": {
        "type": "string",
        "enum": [
          "IDLE",
          "RUNNING",
          "PAUSED",
          "COMPLETED",
          "AWAITING_INTERVENTION",
          "FAILED",
          "CANCELLED"
        ],
        "title": "PipelineRunStatus"
      },
...
      "ProjectOverviewSummary": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "specName": {
            "type": "string",
            "title": "Specname"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/LifecycleState"
          },
          "framework": {
            "type": "string",
            "title": "Framework",
            "default": "Java 21 / Spring Boot 3"
          },
          "databaseEngine": {
            "type": "string",
            "title": "Databaseengine",
            "default": "POSTGRESQL"
          },
          "userStoriesCount": {
            "type": "integer",
            "title": "Userstoriescount",
            "default": 0
          },
          "entitiesCount": {
            "type": "integer",
            "title": "Entitiescount",
            "default": 0
          },
          "testsPassed": {
            "type": "boolean",
            "title": "Testspassed",
            "default": false
          },
          "securityAuditVerdict": {
            "type": "string",
            "title": "Securityauditverdict",
            "default": "PENDING"
          },
          "deploymentStatus": {
            "type": "string",
            "title": "Deploymentstatus",
            "default": "IDLE"
          },
          "deploymentUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deploymenturl"
          },
          "pipelineStatus": {
            "$ref": "#/components/schemas/PipelineRunStatus",
            "default": "IDLE"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "specName",
          "lifecycle"
        ],
        "title": "ProjectOverviewSummary"
      },
...
    }
  }
}